requirements-test.txt
# Test dependencies for integration testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-xdist==3.3.1
requests-mock==1.11.0
moto==4.2.14
boto3-stubs==1.34.0
responses==0.24.1
testcontainers==3.7.1
docker==6.1.3
httpx==0.25.2
uvicorn==0.24.0
fastapi==0.104.1

# For Lambda testing
python-lambda-local==0.1.13
aws-lambda-powertools==2.25.0

# For performance testing
locust==2.17.0

# For mocking external services
wiremock==2.6.0
localstack==3.0.2
